+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/AlertTriggering/MetricAnomalyAlertScope.cs
@@ -43,7 +43,7 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// metric as the scope.
         /// </summary>
         /// <returns>A new <see cref="MetricAnomalyAlertScope"/> instance.</returns>
-        public static MetricAnomalyAlertScope GetScopeForWholeSeries() =>
+        public static MetricAnomalyAlertScope CreateScopeForWholeSeries() =>
             new MetricAnomalyAlertScope(MetricAnomalyAlertScopeType.WholeSeries, seriesGroupInScope: default, topNGroupInScope: default);
 
         /// <summary>
@@ -53,7 +53,7 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// A subset of the possible dimensions of the associated data feed must be set.</param>
         /// <returns>A new <see cref="MetricAnomalyAlertScope"/> instance.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="seriesGroupKey"/> is null.</exception>
-        public static MetricAnomalyAlertScope GetScopeForSeriesGroup(DimensionKey seriesGroupKey)
+        public static MetricAnomalyAlertScope CreateScopeForSeriesGroup(DimensionKey seriesGroupKey)
         {
             Argument.AssertNotNull(seriesGroupKey, nameof(seriesGroupKey));
 
@@ -63,11 +63,13 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// <summary>
         /// Creates a <see cref="MetricAnomalyAlertScope"/> instance in which alerts will only be triggered for anomalies in the top N series.
         /// </summary>
-        /// <param name="topNGroup">Specify the number of timestamps to take into account, and how many anomalies must be in them to send the alert.</param>
-        /// <exception cref="ArgumentNullException"><paramref name="topNGroup"/> is null.</exception>
-        public static MetricAnomalyAlertScope GetScopeForTopNGroup(TopNGroupScope topNGroup)
+        /// <param name="top">The number of timestamps to take into account when choosing the top N series.</param>
+        /// <param name="period">The number of items a period contains.</param>
+        /// <param name="minimumTopCount">The number of anomalies that must be in the specified <paramref name="top"/> number of timestamps to send an alert.</param>
+        /// <returns>A new <see cref="MetricAnomalyAlertScope"/> instance.</returns>
+        public static MetricAnomalyAlertScope CreateScopeForTopNGroup(int top, int period, int minimumTopCount)
         {
-            Argument.AssertNotNull(topNGroup, nameof(topNGroup));
+            var topNGroup = new TopNGroupScope(top, period, minimumTopCount);
 
             return new MetricAnomalyAlertScope(MetricAnomalyAlertScopeType.TopN, seriesGroupInScope: default, topNGroup);
         }